+++ Structure_oM/Reinforcement/IBarReinforcement.cs
@@ -22,43 +22,29 @@
 
 using BH.oM.Base;
 using BH.oM.Structure.MaterialFragments;
-using BH.oM.Spatial.Layouts;
 using BH.oM.Quantities.Attributes;
 using System.ComponentModel;
-using BH.oM.Geometry;
-using System.Collections.Generic;
 
-namespace BH.oM.Structure.SectionProperties.Reinforcement
+namespace BH.oM.Structure.Reinforcement
 {
-    [Description("Defines any transverse reinforcement along a Bar.")]
-    public class TransverseReinforcement : BHoMObject, IBarReinforcement
+    [Description("Base interface for any reinforcement within a BarRebarIntent.")]
+    public interface IBarReinforcement : IBHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        [Description("Layout controlling the reinforcement shape in relation to the bar's section.")]
-        public virtual ICurveLayout CenterlineLayout { get; set; }
-
         [Length]
         [Description("Diameter of a single rebar.")]
-        public virtual double Diameter { get; set; }
+        double Diameter { get; set; }
 
-        public virtual IMaterialFragment Material { get; set; }
-
-        [Length]
-        [Description("Axial distance between consecutive rebars. If AdjustSpacingToFit is true it will become maximum spacing.")]
-        public virtual double Spacing { get; set; }
-
-        [Description("Toggles if provided spacing value should be fixed or adjusted to fit rebars from the host bar's start to end.")]
-        public virtual  bool AdjustSpacingToFit { get; set; }
+        IMaterialFragment Material { get; set; }
 
         [Description("Normalised length (0 means start, 1 means end) along the element where the rebars start.")]
-        public virtual double StartLocation { get; set; } = 0;
+        double StartLocation { get; set; }
 
         [Description("Normalised length (0 means start, 1 means end) along the element where the rebars ends.")]
-        public virtual double EndLocation { get; set; } = 1;
-
+        double EndLocation { get; set; }
 
         /***************************************************/
     }
